@@ -1,13 +1,14 @@
 import React from 'react';
 import { Modal, Button } from 'react-bootstrap';
+import { BsTrash } from 'react-icons/bs';
 
 import { AnimeModel, Status, Score } from '../models';
-import { AnimeModalForm } from './AnimeModalForm';
+import { AnimeModalForm } from './AnimeModalForm/AnimeModalForm';
 
 type Props = {
   editable: boolean;
   onChange: (
-    value: string | undefined | null | Status | Score,
+    value: string | undefined | boolean | null | Status | Score | string[],
     propname: string
   ) => void;
   onDelete: () => void;
@@ -41,7 +42,9 @@ export const AnimeModalItem: React.FC<Props> = ({
       </Modal.Body>
       {editable && <Modal.Footer>
         <Button variant='danger' onClick={onDelete} title='Delete this anime' disabled={!editable}>
-          DELETE
+          <BsTrash />
+          {' '}
+          Delete
         </Button>
       </Modal.Footer>}
     </Modal>
