@@ -1,12 +1,12 @@
 import React from 'react';
 
 import { AnimeModel } from '../models';
-import { AnimeListItem } from './AnimeListItem';
+import { AnimeListItem } from './AnimeListItem/AnimeListItem';
 import { SearchField } from './SearchField';
 
 export type AnimeListProps = {
   setSearch: (newSearch: string) => void,
-  externalAnime: AnimeModel;
+  externalAnime: AnimeModel | null;
   className: string;
 };
 
@@ -14,7 +14,7 @@ export const ExternalAnime: React.FC<AnimeListProps> = ({ setSearch, externalAni
   return (
     <div className={className}>
       <SearchField setValue={setSearch} />
-      {externalAnime && <AnimeListItem editable={true} key={externalAnime.id} {...externalAnime} onChange={() => {}} onDelete={() => {}}/> }
+      {externalAnime && <AnimeListItem editable={true} key={externalAnime.id} {...externalAnime} onChange={() => console.debug('Attempted & ignored change')} onDelete={() => console.debug('Attempted & ignored deletion')}/> }
     </div>
   );
 };
