@@ -4,7 +4,7 @@ import { Spinner } from 'react-bootstrap';
 export const Loading: React.FC = () => {
   return (
     <Spinner animation='border' variant='primary' role='status'>
-      <span className='sr-only'>Loading...</span>
+      <span className='visually-hidden'>Loading...</span>
     </Spinner>
   );
 };
