@@ -1,14 +1,15 @@
 import React, { FormEvent } from 'react';
 import { Button, Form } from 'react-bootstrap';
+import { IoIosAddCircleOutline } from 'react-icons/io';
 
 import { AnimeModel } from '../models';
 import { ExternalAnime } from './ExternalAnime';
-import Accordionned from './Accordionned';
+import { Accordionned } from './Accordionned/Accordionned';
 
 type Props = {
   onSubmit: (a: AnimeModel) => void,
   setSearch: (newSearch: string) => void,
-  externalAnime: AnimeModel
+  externalAnime: AnimeModel | null
 };
 
 export const NewAnimeForm: React.FC<Props> = ( { onSubmit, setSearch, externalAnime }) => {
@@ -20,10 +21,16 @@ export const NewAnimeForm: React.FC<Props> = ( { onSubmit, setSearch, externalAn
 
   return (
     <Accordionned label='Add New Anime'>
-      <Form inline autoComplete='off' onSubmit={handleSubmit}>
-        <Form.Label srOnly htmlFor='name_english'>Add New Anime</Form.Label>
-        <ExternalAnime setSearch={setSearch} externalAnime={externalAnime} className='flex-grow-1 mb-2 mr-sm-2' />
-        <Button type='submit' className='mb-2 mr-sm-2'>ADD ANIME</Button>
+      <Form autoComplete='off' onSubmit={handleSubmit}>
+        <Form.Label visuallyHidden htmlFor='name_english'>Add New Anime</Form.Label>
+        
+        <ExternalAnime setSearch={setSearch} externalAnime={externalAnime} className='flex-grow-1 mb-2 ms-sm-2' />
+        
+        <Button type='submit' className='mb-2 ms-sm-2'>
+          <IoIosAddCircleOutline />
+          {' '}
+          Add
+        </Button>
       </Form>
     </Accordionned>
   );
