@@ -1,10 +1,12 @@
 import React, { FunctionComponent, useMemo } from 'react';
+import { doc, updateDoc, deleteDoc } from 'firebase/firestore';
+
+import { db } from '../firebase';
 import { AnimeModel, Score, Status } from '../models';
-import firebase from '../firebase/firebase';
 
 type ComponentProps = {
   editable: boolean;
-  onChange: (value: string | undefined | null | Status | Score, propname: string) => void;
+  onChange: (value: string | undefined | boolean | null | Status | Score | string[], propname: string) => void;
   onDelete: () => void;
 } & AnimeModel;
 
@@ -15,22 +17,19 @@ type Props = {
 
 export const Anime: React.FC<Props> = ({ component, id, editable, ...rest }) => {
   const docRef = useMemo(() => 
-    firebase
-      .firestore()
-      .collection('animes')
-      .doc(id),
+    doc(db, 'animes', id),
     [id]
   );
   const onChange = (
-    value: string | undefined | null | Status | Score | string[],
+    value: string | undefined | boolean | null | Status | Score | string[],
     propname: string
   ) => {
     if (editable)
-      docRef.update({ [propname]: value });
+      updateDoc(docRef, { [propname]: value });
   };
   const onDelete = () => {
     if (editable)
-      docRef.delete();
+      deleteDoc(docRef);
   };
 
   return React.createElement(component, { id, editable, onChange, onDelete, ...rest });
