@@ -1,30 +1,10 @@
-import React, { useEffect, useReducer, FunctionComponent } from 'react';
+import React, { useEffect, FunctionComponent } from 'react';
+import { collection, query, where, onSnapshot } from 'firebase/firestore';
 
+import { db } from '../firebase';
 import { AnimeModel } from '../models';
-import firebase from '../firebase/firebase';
 import { AnimeListProps } from '../components/AnimeList';
-
-type State = {
-  animes: Array<AnimeModel>;
-};
-
-type Action = {
-  payload: AnimeModel;
-  type: string;
-};
-
-const reducer = (state: State, action: Action) => {
-  switch(action.type) {
-    case 'added': return { ...state, animes: [...state.animes, action.payload] };
-    case 'modified': return { ...state, animes: state.animes.map(a => a.id !== action.payload.id ? a : action.payload) };
-    case 'removed': return { ...state, animes: state.animes.filter(a => a.id !== action.payload.id) };
-    default: return state;
-  }
-};
-
-const INITIAL_STTE: State = {
-  animes: new Array<AnimeModel>()
-}; 
+import useAnimesReducer from '../animeReducer';
 
 type Props = {
   component: FunctionComponent<AnimeListProps>;
@@ -32,20 +12,20 @@ type Props = {
 };
 
 export const PublicAnimes: React.FC<Props> = ({ component, userid }) => {
-  const [state, dispatch] = useReducer(reducer, INITIAL_STTE);
+  const [state, dispatch] = useAnimesReducer();
 
   useEffect(() => {
-    const unsubscribe = firebase
-      .firestore()
-      .collection('animes')
-      .where('uid', '==', userid)
-      .where('isPublic', '==', true)
-      .onSnapshot((snapshot) => {
-        for (const { doc, type } of snapshot.docChanges()) {
-          const payload = { id: doc.id, ...doc.data() } as AnimeModel;
-          dispatch({ type, payload });
-        }
-      });
+    const unsubscribe = onSnapshot(
+      query(
+        collection(db, 'animes'), 
+        where('uid', '==', userid), 
+        where('isPublic', '==', true)
+      ), (snapshot) => {
+        snapshot.docChanges().forEach((change) => {
+          const payload = { id: change.doc.id, ...change.doc.data() } as AnimeModel;
+          dispatch({ type: change.type, payload });
+        });
+    });
 
     return unsubscribe;
   }, [userid]);
