@@ -2,25 +2,28 @@ import { fullAnimeAPI, animeFullAnimeAPIUri, searchFullAnimeAPIUri, streamingUrl
 import { AnimeModel, Score } from '../models';
 import { APISearchResponse, LinkAPIResult, APIAnimeResult } from './anime.models';
 
-function selectUrlInContent(result: LinkAPIResult): string {
+function selectUrlInContent(result: LinkAPIResult): string | null {
   if (!result || !result.content) return null;
+
   const matches = result && result.content && /href=""?([a-zA-Z\\\-/]+)"?"/.exec(result.content);
   if (matches && matches.length >= 2) {
     const uri = matches[1].replace(/\\/g, '');
-    return `https://www2.gogoanime.ee/${uri}`;
+    return `https://anitaku.to/${uri}`;
   }
   return null;
 }
 
-async function getAnimeProfileLink(url: string): Promise<LinkAPIResult> {
+async function getAnimeProfileLink(url: string): Promise<LinkAPIResult | null> {
   try {
     const result = await fetch(url);
     return result.json();
   } catch (err) {
     console.error(err);
+    return null;
   }
 }
 
+// eslint-disable-next-line
 async function getData(url: string): Promise<any> {
   try {
     const result = await fetch(url);
@@ -39,22 +42,23 @@ async function getSearchData(url: string): Promise<APISearchResponse> {
   return getData(url);
 }
 
-export async function searchAsync(query: string): Promise<AnimeModel> {
-  if (!query) return;
+export async function searchAsync(query: string): Promise<AnimeModel | null> {
+  if (!query) return null;
 
-  let animeData: APIAnimeResult;
-  let animeProfileLink: string;
+  let animeData: APIAnimeResult | null = null;
+  let animeProfileLink: string | null = null;
 
   const animeSearchData = await getSearchData(`${fullAnimeAPI}${searchFullAnimeAPIUri}${encodeURIComponent(query)}`);
   if (animeSearchData && animeSearchData.data && animeSearchData.data[0]) {
     animeData = await getAnimeData(`${fullAnimeAPI}${animeFullAnimeAPIUri}/${animeSearchData.data[0].mal_id}`);
+
+    if (!animeData || !animeData.title) return null;
+
     const animeProfileResult = await getAnimeProfileLink(`${streamingUrlAnimeAPI}${encodeURIComponent(animeData.title)}`);
-    animeProfileLink = selectUrlInContent(animeProfileResult);
+    animeProfileLink = animeProfileResult && selectUrlInContent(animeProfileResult);
   }
 
-  if (!animeData) {
-    return null;
-  }
+  if (!animeData || !animeProfileLink) return null;
 
   return {
     name_english: animeData.title_english,
